[PartnerArtist] Avoid circular dependencies
